Separate CurrentNode component from BoardSet
