+++ src/components/BoardSet.css
@@ -10,15 +10,6 @@
   align-self: flex-end;
 }
 
-.comment {
-  width: 300px;
-}
-
-.buttons button {
-  width: 60px;
-  font-size: 120%;
-}
-
 div.mochi div.mochimain span.mochigoma div {
   display: inline-block;
 }
