+++ src/components/BoardSet.js
@@ -10,12 +10,9 @@ Piece.DecoratedComponent.prototype.getPieceImage = PieceHand.DecoratedComponent.
   return `/images/shogi-pieces/${!kind ? "blank.gif" : color + kind + ".svg"}`;
 };
 
-import ForkList from './ForkList';
-
 class BoardSet extends Component {
   render() {
-    const { player, reversed, jumpMap, currentNode, currentPath,
-      previousPath, nextPath, previousForkPath, nextForkPath } = this.props;
+    const { player, reversed } = this.props;
     const playerState = player.getState();
     const players = [
       player.kifu.header["先手"] || player.kifu.header["下手"] || "先手",
@@ -50,37 +47,6 @@ class BoardSet extends Component {
               playerName={players[reversed ? 1 : 0]}
               onInputMove={e => { this.props.onInputMove(e) }}
               reversed={reversed} />
-          </div>
-          <div>
-            <div>
-              {(currentNode.tesuu === 0 ? null : currentNode.tesuu + "手目")} {currentNode.readableKifu}
-            </div>
-            <textarea
-              rows="10"
-              className="comment"
-              placeholder="ここに現在の手についてコメントを書けます。"
-              onChange={e => { this.props.onChangeComments(e.target.value); }}
-              value={currentNode.comment}></textarea>
-            <div className="buttons">
-              <button
-                onClick={e => this.props.onClickPath(previousForkPath)}
-                title="1つ前の分岐に戻る">&laquo;</button>
-              <button
-                onClick={e => this.props.onClickPath(previousPath)}
-                title="1手戻る">&lt;</button>
-              <button
-                onClick={e => this.props.onClickPath(nextPath)}
-                title="1手進む">&gt;</button>
-              <button
-                onClick={e => this.props.onClickPath(nextForkPath)}
-                title="1つ先の分岐に進む">&raquo;</button>
-            </div>
-            <ForkList
-              currentNode={currentNode}
-              currentPath={currentPath}
-              jumpMap={jumpMap}
-              onClickForward={this.props.onClickForward}
-              onClickPath={this.props.onClickPath} />
           </div>
         </div>
       </div>
