+++ src/components/CurrentNode.css
@@ -1,0 +1,8 @@
+.comment {
+  width: 300px;
+}
+
+.buttons button {
+  width: 60px;
+  font-size: 120%;
+}
